feat(FeatureDependency): 댓글 바텀시트 액션의 컨텐츠 구현 (#77)
